MF-91 Added FinalIK to the new IKService
Included the ComputeIK in the new FinalIKService. Using the GameObject of the NewAvatar will get the avatar to reach correctly in the TargetScene, but causes it to float and distort. Additionally, the AvatarScaling Scene does not seem to scale the GameObject. The Prefab on the other will be scaled, but causes some unwanted joint rotations in the TargetScene.